blog_candyy/requirements.txt
attrs==19.1.0
Automat==0.7.0
constantly==15.1.0
cssselect==1.0.3
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.3
Markdown==3.1
parsel==1.5.1
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pytz==2019.1
queuelib==1.5.0
six==1.12.0
sqlparse==0.3.0
w3lib==1.20.0
zope.interface==4.6.0
